fix webpack dev mode
