+++ intuita-webview/package.json
@@ -20,10 +20,10 @@
 		"web-vitals": "^2.1.4"
 	},
 	"scripts": {
-		"start": "react-scripts start",
-		"build": "node ./scripts/build-react-no-split.ts",
-		"test": "react-scripts test",
-		"eject": "react-scripts eject"
+		"start:main": "TARGET_APP=main craco start",
+		"start:sourceControl": "TARGET_APP=sourceControl craco start",
+		"build": "craco build",
+		"test": "craco test"
 	},
 	"eslintConfig": {
 		"extends": [
@@ -44,6 +44,8 @@
 		]
 	},
 	"devDependencies": {
+		"@craco/craco": "^7.1.0",
+		"@craco/types": "^7.1.0",
 		"@types/react-treeview": "^0.4.3",
 		"@types/testing-library__jest-dom": "^5.14.5",
 		"@types/vscode-webview": "^1.57.1",
